chore: Rename Preferences to Settings
