+++ site/src/components/SettingsAccountForm/SettingsAccountForm.tsx
@@ -17,7 +17,7 @@ export const Language = {
   emailLabel: "Email",
   emailInvalid: "Please enter a valid email address.",
   emailRequired: "Please enter an email address.",
-  updatePreferences: "Update preferences",
+  updateSettings: "Update settings",
 }
 
 const validationSchema = Yup.object({
@@ -73,7 +73,7 @@ export const AccountForm: React.FC<AccountFormProps> = ({
 
           <div>
             <LoadingButton color="primary" loading={isLoading} type="submit" variant="contained">
-              {isLoading ? "" : Language.updatePreferences}
+              {isLoading ? "" : Language.updateSettings}
             </LoadingButton>
           </div>
         </Stack>
